WeeklyStatus all with summary flags
For each individual and week combination, 5 columns record whether or not that individual was in one of various categories, such as on the live register, or in education or training.

In order to run the code successfully, it is likely that it will have to be run in stages, rather than all the way through. For that reason, there are various points where it is suggested that the kernel is restarted and the output cleared before proceeding.

This code does not create a column for each status. However, there is no loss of information as the combined statuses are recorded in one column.

The output of the code is a single gzipped CSV file of the DataFrame df3, which can then be imported for further processing and analysis in a new notebook with the code:

from datetime import datetime
today = str(datetime.now()).split('.')[0][0:10]

with gzip.open('\\\\cskma0294\\F\\Evaluations\\JobPath\\WeeklyStatus_flags_'+today+'.gz') as f:
    df = pd.read_csv(f).iloc[:,1:] # drop 1st column (an artefact of the zipping process)

That is, if everything is run on the same day. Otherwise, use the actual file name in place of *** below:

with gzip.open('\\\\cskma0294\\F\\Evaluations\\JobPath\\***') as f:
    df = pd.read_csv(f).iloc[:,1:] # drop 1st column (an artefact of the zipping process)

Alternatively, one could continue working within the notebook with the df3 DataFrame.